code_review_env/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
gradio>=4.0.0
pyyaml>=6.0
